+++ src/components/loading/loading_spinner.test.tsx
@@ -2,7 +2,11 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test/required_props';
 
-import { EuiLoadingSpinner, SIZES } from './loading_spinner';
+import {
+  EuiLoadingSpinner,
+  SIZES,
+  EuiLoadingSpinnerComponentType,
+} from './loading_spinner';
 
 describe('EuiLoadingSpinner', () => {
   test('is rendered', () => {
@@ -17,6 +21,30 @@ describe('EuiLoadingSpinner', () => {
         const component = render(<EuiLoadingSpinner size={size} />);
 
         expect(component).toMatchSnapshot();
+      });
+    });
+  });
+
+  describe('component', () => {
+    (['div', 'span', 'figure'] as EuiLoadingSpinnerComponentType[]).forEach(
+      value => {
+        test(`${value} is rendered`, () => {
+          const component = render(<EuiLoadingSpinner component={value} />);
+
+          expect(component).toMatchSnapshot();
+        });
+      }
+    );
+
+    ['h2'].forEach(value => {
+      test(`${value} is not rendered`, () => {
+        expect(() =>
+          render(
+            // intentionally passing an invalid value
+            // @ts-ignore
+            <EuiLoadingSpinner component={value} />
+          )
+        ).toThrow();
       });
     });
   });