+++ src/components/loading/loading_spinner.tsx
@@ -9,21 +9,26 @@ const sizeToClassNameMap = {
   xl: 'euiLoadingSpinner--xLarge',
 };
 
+export type EuiLoadingSpinnerComponentType = 'div' | 'span' | 'figure';
+
+interface EuiLoadingSpinnerComponentProps {
+  size?: EuiLoadingSpinnerSize;
+  component?: EuiLoadingSpinnerComponentType;
+}
+
 export const SIZES = keysOf(sizeToClassNameMap);
 
 export type EuiLoadingSpinnerSize = keyof typeof sizeToClassNameMap;
-
 export const EuiLoadingSpinner: FunctionComponent<
   CommonProps &
-    HTMLAttributes<HTMLDivElement> & {
-      size?: EuiLoadingSpinnerSize;
-    }
-> = ({ size = 'm', className, ...rest }) => {
+    HTMLAttributes<HTMLDivElement | HTMLSpanElement> &
+    EuiLoadingSpinnerComponentProps
+> = ({ size = 'm', className, component: Component = 'div', ...rest }) => {
   const classes = classNames(
     'euiLoadingSpinner',
     sizeToClassNameMap[size],
     className
   );
 
-  return <div className={classes} {...rest} />;
+  return <Component className={classes} {...rest} />;
 };